+++ tests/reactnative/e2e/app.test.js
@@ -1,12 +1,8 @@
 const jestExpect = require("expect").default;
 
-describe("Web5 Tests", () => {
+describe("App Tests", () => {
   beforeAll(async () => {
     await device.launchApp({ newInstance: true });
-  });
-
-  beforeEach(async () => {
-    await device.reloadReactNative();
   });
 
   it("should have the Web5 test as successful", async () => {
@@ -18,5 +14,16 @@ describe("Web5 Tests", () => {
 
     const web5TestResult = JSON.parse(text);
     jestExpect(web5TestResult.success).toBe(true);
+  });
+
+  it("should have the DWN test as successful", async () => {
+    const dwnTestResultElement = element(by.id("dwn-tests-result"));
+    await expect(dwnTestResultElement).toBeVisible();
+
+    const { text } = await dwnTestResultElement.getAttributes();
+    console.info(">>> DWN Test Results:", text);
+
+    const dwnTestResult = JSON.parse(text);
+    jestExpect(dwnTestResult.success).toBe(true);
   });
 });
